@@ -1,3 +1,3 @@
 module github.com/actforgood/xdi
 
-go 1.16
+go 1.23
